yolov5-farm-cattle/yolov5-farm-cattle/requirements.txt
gradio==4.29.0
numpy
Pillow
torch
matplotlib
opencv-python
